fix: move coupon redeem helper out of component
refs #430
