+++ apps/extension/src/app/components/CouponShopPanel.tsx
@@ -5,6 +5,7 @@ import { demoCouponMetas, type DemoCouponMeta } from '../../extension/couponCata
 import type { CouponRedeemResult } from '../../extension/types'
 import { hudPanelBackground } from '../theme/backgrounds'
 import { renderCouponRedeemStatus } from './couponRedeemStatus'
+import { redeemCouponForPanel } from './redeemCouponForPanel'
 
 const couponMetas: DemoCouponMeta[] = demoCouponMetas
 
@@ -14,52 +15,6 @@ interface CouponShopPanelProps {
   redeemedCouponIds: string[]
   voucherCodes: Record<string, string>
   onRedeem: (couponId: string, cost: number) => Promise<CouponRedeemResult | 'error'>
-}
-
-interface RedeemPanelMessages {
-  alreadyRedeemed: string
-  insufficientBalance: string
-  genericError: string
-}
-
-export async function redeemCouponForPanel({
-  couponId,
-  cost,
-  isAlreadyRedeemed,
-  messages,
-  onRedeem,
-  setError,
-}: {
-  couponId: string
-  cost: number
-  isAlreadyRedeemed: boolean
-  messages: RedeemPanelMessages
-  onRedeem: (couponId: string, cost: number) => Promise<CouponRedeemResult | 'error'>
-  setError: (message: string) => void
-}): Promise<void> {
-  if (isAlreadyRedeemed) {
-    setError(messages.alreadyRedeemed)
-    return
-  }
-
-  try {
-    const result = await onRedeem(couponId, cost)
-    if (result === 'already_redeemed') {
-      setError(messages.alreadyRedeemed)
-      return
-    }
-    if (result === 'insufficient') {
-      setError(messages.insufficientBalance)
-      return
-    }
-    if (result === 'error') {
-      setError(messages.genericError)
-      return
-    }
-    setError('')
-  } catch (err) {
-    setError(err instanceof Error && err.message ? err.message : messages.genericError)
-  }
 }
 
 export function CouponShopPanel({
@@ -98,7 +53,6 @@ export function CouponShopPanel({
       await redeemCouponForPanel({
         couponId: selectedCoupon.id,
         cost: selectedCoupon.price,
-        isAlreadyRedeemed: false,
         messages: {
           alreadyRedeemed: t('coupon.alreadyRedeemed'),
           insufficientBalance: t('coupon.insufficientBalance'),
